+++ src/GcpMetadata/Error.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace Kreait;
+namespace Kreait\GcpMetadata;
 
 class Error extends \RuntimeException
 {
-
 }